Changes to data management
The CSV Parser now no longer uses its own player vector, but instead accesses and writes to the global manager and its player vector. This unifies the data structure and makes it easier to manage when merging with the data coming from the XML. The global manager has been extended to fully support that role.